+++ liteconfig_py/__init__.py
@@ -13,4 +13,4 @@ try:
 except ImportError:
     __all__ = ['Config']
 
-__version__: str = '0.2.0'
+__version__: str = '0.3.0'